Fixed btsetcharvalue() and btgetcharvalue() crashers. The problem was as follows: (from btechstats.c)
This function was being used for parsing object references:

dbref char_lookupplayer(dbref player, dbref cause, int key, char *arg1)
{
    dbref which;

    if (!arg1 || !*arg1)
        return NOTHING;

    if (!string_compare(arg1, "me") && (Typeof(player) == TYPE_PLAYER))
        return player;
    if (arg1[0] == '#') {
        if (sscanf(arg1, "#%d", &which) == 1)
            if (which >= 0 && isPlayer(which))
            return which;
        return NOTHING;
    }
    return lookup_player(NOTHING, arg1, 0);
}

There are two problems with it, first of, it's reinventing the wheel. The mux
already has a lookup_player() function to do this.

Second, it doesn't verify the validity of the input, in this case, isPlayer(arg)
where arg was larger than the db would cause a segmentation fault. You should use
isGood(x) to check a dbref before using it.

-Hag