This is a semi-critical patch to get weather to work properly. (I.e. Vis)
Weather randomly updates currently.  For some reason, when map_update gets called, the muxevent_tick it grabs is odd, sometimes it is even.
This causes problems with the checking part of the weather, as it checks muxevent_tick % 60, which will never be 0 when its odd.

I tested several times and could get it to come up odd or even randomly when i @restarted.

The simple fix is to make it % on a number that comes up under odd/even, which, I arbitrarly picked 25, which makes weather checks every 50 seconds versus every 60 in 
the old code.

Anyways, this is a long commit message for a 2 character code fix.